Moving working Go version to here
I have a working Go implementation of this project and I originally had it in my catch-all personal projects repository. I am now moving it over to this repo as it should have been in the first place. This repo was originally created because I was doing a full rewrite in Rust. This repo will now hold the Go code and the Rust code as I work on the migration, if it ever even gets finished. I might just stick with Go, we shall see.