Rust posttemp for Scylla functional
Successfully implemented the posttemp route, but only for Scylla. This progress was made before I decided to switch all of my projects over to using sqlite. The route also creates an entirely new connection to the database every time posttemp is called. This is obviously not the best, but I could not figure out how to do it properly at the time, so that was a termporary fix.